chore: move operations/types to UserService if not admin-scoped
