--- ui/user/src/lib/components/admin/filters-drawer/types.ts
@@ -1,3 +1,0 @@
-import type { AdminService } from '$lib/services';
-
-export type FilterOptionsEndpoint = typeof AdminService.listAuditLogFilterOptions;
