+++ ui/user/src/routes/admin/agents/+page.ts
@@ -1,6 +1,5 @@
 import { handleRouteError } from '$lib/errors';
-import { AdminService, NanobotService } from '$lib/services';
-import type { OrgUser } from '$lib/services/admin/types';
+import { UserService, NanobotService, type OrgUser } from '$lib/services';
 import type { ProjectV2Agent } from '$lib/services/nanobot/types';
 import type { PageLoad } from './$types';
 
@@ -11,7 +10,7 @@ export const load: PageLoad = async ({ fetch, parent }) => {
 	try {
 		[agents, users] = await Promise.all([
 			NanobotService.listAllNanobotAgents({ fetch }),
-			AdminService.listUsers({ fetch })
+			UserService.listUsers({ fetch })
 		]);
 	} catch (err) {
 		handleRouteError(err, `/agents`, profile);
