+++ ui/user/src/routes/admin/groups/types.ts
@@ -1,4 +1,4 @@
-import type { OrgGroup, GroupRoleAssignment } from '$lib/services/admin/types';
+import { type OrgGroup, type GroupRoleAssignment } from '$lib/services';
 
 export interface GroupAssignment {
 	group: OrgGroup;
