+++ ui/user/src/routes/admin/mcp-servers/c/[id]/instance/[ms_id]/details/+page.svelte
@@ -5,12 +5,7 @@
 	import McpServerActions from '$lib/components/mcp/McpServerActions.svelte';
 	import OAuthMetadataDebug from '$lib/components/mcp/OAuthMetadataDebug.svelte';
 	import { DEFAULT_MCP_CATALOG_ID, PAGE_TRANSITION_DURATION } from '$lib/constants';
-	import {
-		AdminService,
-		UserService,
-		type MCPCatalogServer,
-		type OrgUser
-	} from '$lib/services/index.js';
+	import { UserService, type MCPCatalogServer, type OrgUser } from '$lib/services/index.js';
 	import { profile } from '$lib/stores/index.js';
 	import { Info } from 'lucide-svelte';
 	import { fly } from 'svelte/transition';
@@ -33,7 +28,7 @@
 		compositeParentName = mcpServer.compositeName;
 
 		if (mcpServer.userID && !isSameUser) {
-			const user = await AdminService.getUser(mcpServer.userID);
+			const user = await UserService.getUser(mcpServer.userID);
 			connectedUsers = [user];
 		}
 	}
