+++ ui/user/src/routes/admin/users/+page.ts
@@ -1,13 +1,12 @@
 import { handleRouteError } from '$lib/errors';
-import { AdminService } from '$lib/services';
-import type { OrgUser } from '$lib/services/admin/types';
+import { UserService, type OrgUser } from '$lib/services';
 import { profile } from '$lib/stores';
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async ({ fetch }) => {
 	let users: OrgUser[] = [];
 	try {
-		users = await AdminService.listUsers({ fetch });
+		users = await UserService.listUsers({ fetch });
 	} catch (err) {
 		handleRouteError(err, `/users`, profile.current);
 	}
