+++ ui/user/src/routes/mcp-servers/s/[id]/details/+page.svelte
@@ -5,7 +5,7 @@
 	import OAuthMetadataDebug from '$lib/components/mcp/OAuthMetadataDebug.svelte';
 	import { PAGE_TRANSITION_DURATION } from '$lib/constants';
 	import Loading from '$lib/icons/Loading.svelte';
-	import { AdminService, UserService, type MCPServerInstance, type OrgUser } from '$lib/services';
+	import { UserService, type MCPServerInstance, type OrgUser } from '$lib/services';
 	import { profile } from '$lib/stores';
 	import { onMount } from 'svelte';
 	import { fly } from 'svelte/transition';
@@ -29,7 +29,7 @@
 				mcpServer.id
 			);
 		}
-		users = await AdminService.listUsersIncludeDeleted();
+		users = await UserService.listUsersIncludeDeleted();
 		loading = false;
 	});
 	let title = $derived(mcpServer?.manifest.name);
